rename components and set nrc content as container
