+++ src/containers/nrc-content/nrc-challenges/nrc-challenges-constants.js
