+++ src/containers/nrc-content/nrc-challenges/nrc-challenges-selectors.js
