+++ src/containers/nrc-content/nrc-content.js
