+++ src/containers/nrc-content/nrc-footer/nrc-footer-component.jsx
