+++ src/containers/nrc-content/nrc-indicators/indicator-card/indicator-card.js
