+++ src/containers/nrc-content/nrc-indicators/nrc-indicators-selectors.js
