+++ src/containers/nrc-content/nrc-indicators/nrc-indicators-styles.module.scss
