+++ src/containers/nrc-content/nrc-trend/nrc-trend.js
